api/requirements.txt
boto3==1.42.37
strands-agents==1.23.0
fastapi==0.128.0